Add an example of using Dropwizard with Conscrypt
Conscrypt is used as the JCE provider for a HTTP/2 app.
