+++ dropwizard-http2/src/test/java/io/dropwizard/http2/AbstractHttp2Test.java
@@ -2,11 +2,17 @@ package io.dropwizard.http2;
 
 import com.google.common.base.Charsets;
 import io.dropwizard.logging.BootstrapLogging;
+import io.dropwizard.testing.ResourceHelpers;
 import org.eclipse.jetty.client.HttpClient;
 import org.eclipse.jetty.client.api.ContentResponse;
 import org.eclipse.jetty.client.api.Result;
 import org.eclipse.jetty.client.util.BufferingResponseListener;
 import org.eclipse.jetty.http.HttpVersion;
+import org.eclipse.jetty.http2.client.HTTP2Client;
+import org.eclipse.jetty.http2.client.http.HttpClientTransportOverHTTP2;
+import org.eclipse.jetty.util.ssl.SslContextFactory;
+import org.junit.After;
+import org.junit.Before;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
@@ -20,6 +26,24 @@ public class AbstractHttp2Test {
 
     static {
         BootstrapLogging.bootstrap();
+    }
+
+    final SslContextFactory sslContextFactory = new SslContextFactory();
+    HttpClient client;
+
+    @Before
+    public void setUp() throws Exception {
+        sslContextFactory.setTrustStorePath(ResourceHelpers.resourceFilePath("stores/http2_client.jts"));
+        sslContextFactory.setTrustStorePassword("http2_client");
+        sslContextFactory.start();
+
+        client = new HttpClient(new HttpClientTransportOverHTTP2(new HTTP2Client()), sslContextFactory);
+        client.start();
+    }
+
+    @After
+    public void tearDown() throws Exception {
+        client.stop();
     }
 
     protected static void assertResponse(ContentResponse response) {
