+++ dropwizard-http2/src/test/java/io/dropwizard/http2/Http2IntegrationTest.java
@@ -5,14 +5,8 @@ import io.dropwizard.Configuration;
 import io.dropwizard.testing.ConfigOverride;
 import io.dropwizard.testing.ResourceHelpers;
 import io.dropwizard.testing.junit.DropwizardAppRule;
-import org.eclipse.jetty.client.HttpClient;
-import org.eclipse.jetty.http2.client.HTTP2Client;
-import org.eclipse.jetty.http2.client.http.HttpClientTransportOverHTTP2;
-import org.eclipse.jetty.util.ssl.SslContextFactory;
 import org.glassfish.jersey.client.JerseyClient;
 import org.glassfish.jersey.client.JerseyClientBuilder;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -33,24 +27,6 @@ public class Http2IntegrationTest extends AbstractHttp2Test {
             ConfigOverride.config("tls_http2", "server.connector.trustStorePath",
                     ResourceHelpers.resourceFilePath("stores/http2_client.jts"))
     );
-
-    private final SslContextFactory sslContextFactory = new SslContextFactory();
-    private HttpClient client;
-
-    @Before
-    public void setUp() throws Exception {
-        sslContextFactory.setTrustStorePath(ResourceHelpers.resourceFilePath("stores/http2_client.jts"));
-        sslContextFactory.setTrustStorePassword("http2_client");
-        sslContextFactory.start();
-
-        client = new HttpClient(new HttpClientTransportOverHTTP2(new HTTP2Client()), sslContextFactory);
-        client.start();
-    }
-
-    @After
-    public void tearDown() throws Exception {
-        client.stop();
-    }
 
     @Test
     public void testHttp11() throws Exception {
