+++ dropwizard-http2/src/test/java/io/dropwizard/http2/Http2WithConscrypt.java
@@ -2,6 +2,7 @@ package io.dropwizard.http2;
 
 import io.dropwizard.Configuration;
 import io.dropwizard.testing.junit.DropwizardAppRule;
+import org.conscrypt.OpenSSLProvider;
 import org.eclipse.jetty.client.HttpClient;
 import org.eclipse.jetty.http2.client.HTTP2Client;
 import org.eclipse.jetty.http2.client.http.HttpClientTransportOverHTTP2;
@@ -11,18 +12,23 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
+import java.security.Security;
 import java.util.Optional;
 
 import static io.dropwizard.testing.ConfigOverride.config;
 import static io.dropwizard.testing.ResourceHelpers.resourceFilePath;
 
-public class Http2WithCustomCipherTest extends AbstractHttp2Test {
+public class Http2WithConscrypt extends AbstractHttp2Test {
 
-    private static final String PREFIX = "tls_custom_http2";
+    static {
+        Security.addProvider(new OpenSSLProvider());
+    }
+
+    private static final String PREFIX = "tls_conscrypt";
 
     @Rule
     public final DropwizardAppRule<Configuration> appRule = new DropwizardAppRule<>(
-        FakeApplication.class, resourceFilePath("test-http2-with-custom-cipher.yml"),
+        FakeApplication.class, resourceFilePath("test-http2-with-conscrypt.yml"),
         Optional.of(PREFIX),
         config(PREFIX, "server.connector.keyStorePath", resourceFilePath("stores/http2_server.jks")),
         config(PREFIX, "server.connector.trustStorePath", resourceFilePath("stores/http2_client.jts"))
