+++ dropwizard-http2/src/test/resources/test-http2-with-conscrypt.yml
@@ -1,0 +1,11 @@
+server:
+  type: simple
+  connector:
+    type: h2
+    port: 0
+    keyStorePassword: http2_server
+    trustStorePassword: http2_client
+    validateCerts: false
+    jceProvider: Conscrypt
+  applicationContextPath: /api
+  adminContextPath: /admin
